fix(ci): Add int cast to getDailyTransactionCount, fix BaaS PHPStan errors
- Add (int) cast to FraudDetectionService::getDailyTransactionCount()
  return value (MySQL count() returns string, not int)
- Fix null safety in PartnerAuthMiddleware IP check
- Fix BaaSWorkflowTest constructor args after PartnerBillingService refactor
- Add PHPStan baseline for BaaS v2 test type-annotation warnings

+++ app/Domain/Fraud/Services/FraudDetectionService.php
@@ -446,7 +446,7 @@ class FraudDetectionService
      */
     protected function getDailyTransactionCount(User $user): int
     {
-        return Cache::remember(
+        return (int) Cache::remember(
             "user_daily_txn_count_{$user->id}",
             60, // 1 minute cache
             function () use ($user) {
